arm32: Remove ACTLR emulation
It is unclear for what the emulation to the implementation-defined ACTLR
register (Auxiliary Control Register) was required. In general, this
code emulated the access for only 32-bit userland. On 64-bit, the
instructions to read/write the ACTLR register are different (MRS/MSR
instead of MRC/MCR).

Hence, remove the code for emulating this register access.

Change-Id: Ie72e853029843f5bf3c4deec74e4d0e81451103e