ia32: Remove unnecessary context cache flush
Non-present context entries are already flushed when they transition to
present, no need to flush here. Apart from that, a flush here would not
be sufficient anyway, as it only flushes the entry that triggered the
allocation of the context table, but not entries in the same context
table that are requested subsequently.

Change-Id: I9d6ee36bdd44f94194af7f59d9bbfba42af64848