with CONFIG_NO_FRAME_PTR=n, use '-fno-omit-frame-pointer'
The effect of the configuration option CONFIG_NO_FRAME_PTR is to add
'-fomit-frame-pointer' to the compiler parameter which means that the
compiler might have an additional register on certain architectures.

Actually '-fomit-frame-pointer' is enabled by default with -O1 and
higher. Sometimes it is useful for debugging to force the compiler to
generate stack frames. Adding '-fno-omit-frame-pointer' makes it easier
on certain architectures (e.g. AMD64) to walk the kernel to find the
calling functions.

CONFIG_NO_FRAME_PTR is enabled by default, so there is no change in
behavior if this option was not changed.

Change-Id: I0f6d4b77b89843ca16e4d6d0b3d90ae7745e5ca0