Fix object space memory ordering
When invoking an object, the capability lookup is done without taking
any locks. Only updates to the object space are protected by the space
lock.

When new capability slots are allocated, it might be necessary to
allocate a new page, scrub it and make it visible in the object space
data structure. It must be ensured that other CPUs in the system that
traverse the data structure concurrently do not see any stale data of
the page. This requires a write memory barrier to make the memset()
observable before the pointer or page table is updated.

On the read side there is no read memory barrier required. There are
always data dependencies between the loads that all contemporary
architectures honour without any special treatment.

Change-Id: Ieec62962282865884c6552a5f6dd30ae972284b8