arm/gic: consider all INTIDs on GICv3+
On GICv2 bits 12..10 of the INTID are used to encode the source CPU
interface for SGIs and are otherwise RES0. But on GICv3 with affinity
routing enabled, they are treated normally and will be potentially set
in the extended PPIs, extended SPIs and LPIs ranges. While Fiasco only
supports LPIs, the previously used mask would lead to aliasing of LPIs
every 1024 INTIDs.

Change-Id: I8df5019918f30c6f1cf57de20e0f643634081d44