jdb: Rework the handling of the current thread
Remove the global current_active thread pointer and instead always get
the most appropriate thread explicitly in the given context. The use of
current_active was error-prone as it did not take into account the
difference between the current thread on the current CPU and the
current thread on a CPU at hand.

Most notably, this made the display of the properties of the current
thread unusable unless the user has been examining the current thread
on the current CPU.

The Jdb::get_current() method with its side-effect behavior was also
problematic.

This change also enables to streamline the code in other places, e.g.
remove the prompt extension update() and related methods.

Change-Id: I794a43b33396608ae62f8daea37ef739bb5b5159