ARM: Drop DMB instruction in syscall page on ARMv5
The Data Memory Barrier (DMB) instruction was introduced in ARMv6.
Trying to execute it on a ARMv5 CPU results in an undefined instruction
exception, which happens e.g. in ned with newer GCC versions:

  EXCEPTION: (00) undef insn pfa=0107152c, error=00000000 psr=60000010
  L4Re: rom/ned: Unhandled exception: PC=0xffffff44 PFA=0x1071528

It does not look like there is a good replacement for ARMv5: The
stronger Data Synchronization Barrier (DSB) instruction supported in
some ARMv5 processors (back then still called "Drain Write Buffer", DWB)
is not available in user mode (the MCR instruction can only be used in
privileged modes on ARMv5).

Given that these old CPUs are likely usually used in a single core
configuration (and without sophisticated out-of-order execution) just
drop the barrier entirely for ARMv5. This also matches the behavior for
kernel code in Fiasco (Mem::dmb()) which is just a compiler (memory)
barrier on ARMv5.

Change-Id: I500f55e6ce220485b37eede976d8f65944cfb299