jdb: make kobject names independent of jdb
The kernel object debug information is also useful to a JTAG debugger.
It is essential for the Trace32 OS awareness because it's the only way
to enumerate all kernel objects. On the other hand, it is not always
desirable to include the full fledged JDB just to enable JTAG
debugging.

Make this information independent of the JDB frontend and expose just
the debugger capability to set object names.

Change-Id: Ie17514e8ea26c71a1afebb2a000ace88cf8be6a3