arm: add call to set ASID of task
On Cortex-R52 based platforms, the VMID is transmitted on the bus for
device memory accesses. This is used by some SOCs as requester
authentication in their IO-MPU. Allow a privileged component in the
system to explicitly set the VMID to grant access for applications or
virtual machines.

Because this is a very special concept of the R52, the interface is
attached to the platform control capability. It shall not be needed on
any other platforms.

Change-Id: I93058fbb5cddabf7a79069ad9df1baee57d23e83