arm: add proper guest vtimer handling
Originally Fiasco masked the elapsed vtimer interrupt in the CNTV_CTL
register which belongs to the guest. While it worked on Linux it does
not on various other real time guests.

Now if the vtimer is configured for direct injection Fiasco will mask
the vtimer PPI in the GIC instead. The PPI will be unmasked as soon as
the guest has EOIed the virtual vtimer interrupt and we re-enter the
guest.

Change-Id: Iad484026b7911528f577487cb3c46b9857260774