Evade GNU Make bug resulting in outdated globalconfig on Kconfig change
When the Fiasco build system detects that Kconfig files have changed,
it uses the oldconfig mechanism to regenerate globalconfig.out.

After that make should restart itself, because globalconfig.out is
included earlier in the Makefile. But due to a bug in Make that does
not happen, because globalconfig.out and the derived globalconfig.tags
are specified as prerequisites of $(srcdir)/Makeconf, which itself has
an empty recipe assigned, i.e. never changes.
Unfortunately Make had a bug where "if an included makefile was rebuilt
as a prerequisite of another included makefile which didn't need to be
rebuilt, make would not realize that it needed to re-exec itself".
The bug has only been fixed with Make 4.3.90.

With versions prior to 4.3.90, the lack of a restart then results in an
outdated globalconfig.out being used to generate .Module.deps, which can
then later result in build errors.

We work around the make bug by removing the statements that specify
globalconfig.tags and globalconfig.out as prerequisites of
$(srcdir)/Makeconf, since it is not apparent why they should be
prerequisites of it in the first place.

See ticket #CD-844.

Change-Id: I450b76d3c98cc5aa8836e0470bf2314d60fd1a5a