VMX: Fix single-context EPT invalidation
For the single-context EPT invalidation to work correctly, we have to
store the actual EPT Pointer (including its configuration bits) to the
INVEPT descriptor, not just the physical address of the EPT (see the
Intel manual, Section 30.3).

Since we actually have no use for just the physical address of the EPT
in Vm_vmx_ept, we replace the _ept_phys member with _eptp.

Additionally, we indicate that the INVEPT instruction clobbers the
flags (which is how INVEPT indicates success/failure).

See ticket #CD-824.

Change-Id: Iac586d9fcbe05f2c875e4ceb0e580e1804381f5e