arm: trap PMU access on cpu_virt
On arm64 such accesses are already trapped because the respective PMU
registers are not context switched. Do the same on arm32 to keep things
simple.

Note that the affected bits are officially "Reserved, UNK/SBZP" on CPUs
that do not implement the performance monitoring extension. It should
still be safe to set these bits because this will be done only on
Armv7+ CPUs and it is highly unlikely that a conforming implementation
would use the bits for something else.

Change-Id: I980843881e5a770f44e248f2a853438d9f0ff13b