Make the __libc_backend_outs() implementations compliant
The documentation of __libc_backend_outs() in drivers/libc_backend.h
requires that the implementations return 1 on success and 0 otherwise.
This makes the two implementations compliant with the documentation.

The implementation of __libc_backend_outs_wrapper() has been altered to
make the return value consistent with its intended use (despite the fact
that the return value is always ignored in __v_printf()).

This change also makes the reqspec4.kernel.libc.Putchar requirement
testable without altering the implementation of putchar().

Change-Id: I72ea2a270d1f817e3e38d2328beef8ee3fd977c6