riscv: Flush TLB after handling sigma0 page fault
On the HiFive Unmatched board the boot got stuck in sigma0, the cause
was a constantly recurring instruction page fault in sigma0. Performing
a TLB flush after updating the page table resolves the issue.

This is necessary because the RISC-V specification permits the caching
of PTEs whose V (Valid) bit is clear

Change-Id: Id24ccc7f461511ed723c1726d7f072c46d051732