UART checksumming
The purpose of the UART checksumming is to make sure the output of
Fiasco transmitted over a serial line is not accidentally altered, thus
potentially turning a failing test into a passing test, etc.

The UART output of Fiasco is checksummed continously using a CRC32
algorithm and the current checksum is printed periodically with each
newline using a checksum tag. There is a counterpart decoding filter
that checks the consistency of the checksums and removes the checksum
tags from the output.

The checksum tag size is 12 characters and it has the following format:

    "\n{hhhhhhhh} "

In this template, the 'h' characters represent lower-case hexadecimal
characters that encode the CRC32 checksum of all the previous
characters (including newline characters, but excluding the actual
checksum tag, i.e. the curly brackets, the hexadecimal characters in
between them and the tailing character).

The checksumming starts only after the first checksum tag which is:

    "\n{00000000} "

Change-Id: I54484784c2acea1e3dd1089ac0e21e7375783899