arm/arm64: consider FnV for instruction/data aborts
arm64: The ESR.FnV bit defines in certain cases (ESR.{IFSC,DFSC}=0x10)
if the FAR register is valid. Otherwise this bit is RES0.

arm32/!cpu_virt: On ARMv8+, the FnV bit is bit 16 in IFSR/DFSR.

arm32/cpu_virt: On ARMv8+, the FAR register is invalid either if DFSC is
set to 0x11 (this should be only possible if FEAT_RAS is implemented and
only for data aborts) or if the HSR.FnV bit is set.

See ticket #CD-840.

Change-Id: I632867d584f038058a222d00980624b20c75f18a