add support for second uart
Code for a second Uart instance is useful if the platform-provided Uart
is not known at build time.

In that case, add
- `CXXFLAGS_uart-libuart += $(call LIBUART_UART2, foo)`
  (basically very similar to existing `CXXFLAGS_uart-libuart` lines)
- `PREPROCESS_PARTS += second_uart`
- `OBJECTS_LIBUART += uart_foo.o`

to the platform Modules file to add L4::Uart_foo which is declared in
the header file uart_foo.h. By default, the Uart which is declared by
`$(call LIBUART_UART)` is initialized. To use the second Uart instead,
call

  Uart::set_uart_instance(instance)

with `instance > 0`. This has to be done before Uart::startup() is
executed (called from Kernel_uart::init() which is called from
Startup::stage1()).

Change-Id: Ida6eb7abf2bee9a5652e56a83bb78e6f64a9ae5f