Implement a custom has_sibling() for D_list
The D_list cannot inherit the has_sibling() method from D_list_cyclic
because pushing an item to an empty D_list does not use the self-list
of the item (as is the case for Sd_list), but the _h member.

The method also cannot be static, because the _h member is not static
(in theory, _h could be a static member, but the D_list could not be
implemented as a header-only class in that case).

Change-Id: I3793aad05a4c7c08ac2c89ef332f56efb3d915f2