templates: Add build configs for Qualcomm platforms
Add configuration templates for the new Qualcomm platforms so they are
at least build-tested by the buildbot.

It is important to test the Qualcomm platforms with IOMMU/cpu-virt both
on and off because currently special firmware is required to make use of
them, so it is quite likely that some users would configure them without
full functionality.

Change-Id: I1f9e2c0e02ee9ffc7a0721a134a7bc05b598711b