preprocess: Consider comments when replacing EXTENSION
The lines that the EXTENSION statement spanned, which are not
incorporated in the preprocessed code, must be added as an offset
to the line directive to preserve a correct line mapping.
This is important to allow accurate source code annotations in
compiler warning and error messages.

Preprocess discards comments on EXTENSION statements, but does not
adjust the line directive offset correctly because the string matched
against does not contain the comments themselves but references to the
comments array.

Thus, with this change comments are now expanded before calculating the
line directive offset for an EXTENSION statement.

Change-Id: I49da8243681e3305abfcf5413e2a0b6228e7686e