riscv: Rework emulation of periodic timer
The previous approach to kick off the timer in Timer::init() breaks the
initialization of the delay loop.

The reason for that is that IRQs are only opened shortly before
initializing the delay loop during the initialization of the delay loop,
so the timer is lagging massively behind its schedule.
Over time it catches up, but the wrong measurements in Delay::init()
have already happened then.

Instead kick off the timer, i.e. program the next timer interrupt as
current time plus timer period, when it is enabled via Timer::enable().
That is done much later than Timer::init(), immediately before
the initialization of the delay loop.

Change-Id: If820a0b2b04634fe8913c6e948f9e3f6eac8086b