arm: start vCPUs with default cachability
Make the live of user space a bit easier by changing the vCPU user mode
default to set "default cachability". This enables code to run in vCPU
user mode like in vCPU kernel mode. The pingpong application actually
relies on this. A virtual machine monitor will set the HCR to something
different, so real guests are not affected by this change.

Change-Id: I2cba703dbc31df8bec49b2ced6d36b6aa4e680f2