Arm: bsp: qcom: Fix IOMMU on MSM8909
Like on the MSM8916 platform the SMMU on MSM8909 reports support for
cache coherent page table walks. However, it turns out that this is
incorrect: the SMMU is actually not cache coherent and selecting
ARM_IOMMU_COHERENT causes random context faults when the page tables
have not been flushed to RAM yet.

Fix this by dropping the "select ARM_IOMMU_COHERENT". Note that this
change introduces a harmless warning during boot:

  KERNEL: Warning: IOMMU: Configured as non-dma-coherent, but SMMU
  reports that coherent page table walks are supported.

This is essentially the opposite case of the NXP LX2160 added recently:
That one reports no support for coherent page table walks but actually
supports them, and here support is advertised but actually non-existent.

On MSM8916 coherent page table walks work correctly so this is likely a
hardware difference between MSM8916 and MSM8909 (perhaps introduced by
replacing the Cortex-A53 with older Cortex-A7).

Change-Id: I37240793b0b5b378b9ea6b08a17374ebd37c70fe