add test counter and check for increment
Change one reserved field in the KIP to be used as the
test counter for model-based testing. Furthermore, check
if the third user register in the UTCB contains a magic
value during a syscall, and increment this test counter if
it does.

This change also adds a new configuration option MBT_COUNTER.

Change-Id: Ifb35601606f296a4f9efacc3afe059311e5a73d9