semaphore: unblock all waiting threads on destroy
When a semaphore is destroyed, unblock all waiting threads and remove
them from the _waiting queue because this must not be done after
returning from Context::schedule() in Semaphore::sys_down() -- the
Semaphore context might be invalid then.

As the _waiting.lock() is a spinlock, separately grab this lock for
every member of the _waiting list and release it before potentially
switching away.

Also, in down(), check if the existence lock is still valid after
grabbing the _waiting.lock().

See ticket #CD-431.

Change-Id: Ida79520a69b0ddd8b63a2f7e0d3069d7eaaeb3fd