arm: panic on calling non-initialized nested trap handler
If the kernel was configured to access a certain hardware device which
isn't available, the corresponding kernel page fault might be triggered
before the nested trap handler (for calling the kernel debugger) is
initialized. Instead of calling the non-initialized handler, dump the
trap state and panic to have a informative feedback for the developer.

A similar test exists in IA32 (cf. handle_not_nested_trap()).

Change-Id: Id8cbcfb0d7e6b5e3ab7876cad9f0818fd9b945a8