preprocess: Strip off paths from filenames
This allows preprocess to strip off a prefix path
from the file name generated into the #line
directives which makes __FILE__ statements
much smaller in the binary and output.
Use it in Fiasco's Makeconf to strip off
the path to the Fiasco source, and thus
not showing a possible long path to the
source.

Change-Id: I7d9728fc43d282ff75729b074e1bd49da6522eed