abi: fix upper limit of BR register index for abs timeouts
An absolute timeout is encoded in a buffer register, with the index of
the buffer register specified in the timeout value. The kernel has to
crop the index to the number of defined buffer registers.

The code without this fix allows to refer to the 59th buffer register
while the UTCB only defines 58 buffer registers (off-by-one error).
This is not really an issue because in the worst case the user just
shoots into his own feet by using the UTCB error code as absolute
timeout value.

This value is only read by the kernel and the offset is guaranteed to
be inside the UTCB. Therefore, even without this fix, there is no harm
done.

Thanks lifetimeosdev!
See https://github.com/kernkonzept/fiasco/issues/7

Change-Id: Id1b9b1be121672d3e7679efa47b1997985e4cb0e
Requires-Change-Id: I7b32f1fc6072121d07e5446baa780d177eb98288