spin_lock: Document memory order guarantees
The memory order guarantees provided by the Spin_lock were not
specified, which makes it difficult to reason about the validity of
certain constructs.

The memory order behavior that has now been documented should be
sufficient to ensure that Spin_lock behaves as expected, and relaxed
enough to not restrict ourselves unnecessarily with regard to any
optimisation of its implementation (currently seems to be rather
unoptimized, e.g. there are duplicate barriers in some cases).

Change-Id: I74213cc1fd233e9c8808ed587a81fe5ee1613a38