force access_once()/write_now() always inline
With -Os the compiler strictly prevents inlining if this increases the
code size. For these two functions, the inline optimization is done
without knowing the actual size of the assembler statement. At least on
ARM/ARM64, inlining makes definitely sense for these two functions
because a function call has the same instruction size as the generated
load/store instruction.

Change-Id: Id030958de0327fb0255dfc18ed64319e5fb76a46