Add make command to generate compile_commands.json
Although the generated compile_commands.json file comes close, it is
not 100% complete. More specifically, it is missing a few object files
that are not build via the regular OBJS/SRC_ALL mechanism.
The known cases are:
  - Makerules.CHECKSUM
  - Makerules.FONTS
  - Makerules.LIBUART: uart objects file specified via OBJECTS_UART
  - Makerules.TCBOFFSET

Overall, however, these are rare exceptions and the affected files are
not that interesting to begin with, so it should be bearable.

The correctness of the generated compile_commands.json was checked by
comparing it with a compile_commands.json generated by bear.

Change-Id: I8c0929aa3d9368c42939124651b9f62b7854d372