semaphore: Fix race condition in semaphore timeout handling
Using Semaphore::down() with a timeout can easily trigger a race
condition that causes the semaphore counter to become inconsistent.
In particular, it happens when Semaphore::up() is called from another
thread approximately the same time the down() timeout expires.

An ex_regs cancel operation can also trigger this race condition.

To avoid that in case of a race between up() and timeout/cancel the
semaphore counter becomes inconsistent, detect the race and report
the down operation as successful.

See ticket #CD-763.

Change-Id: I4aa01e771452db7f0155c089c33c9f72094f227d