Arm/SMMU-v3: Add KIP feature tag
Add the feature tag "arm,smmu-v3" such that bootstrap
can know whether the platform must have a SMMU-v3 available such that
the boot is successful. In case Fiasco is configured to use SMMU-v3 but
the hardware does not have it, such as possible with QEMU, the boot-up
crashes. The way to detect the availability is the device tree available
in bootstrap.

Change-Id: I3c7f89c575faa240aa54cbe34b0f2f2913980323