arm: always run threads in EL0 by default
On a !cpu_virt kernel, threads naturally run in EL0. We now keep the
same default for cpu_virt kernels too. Because a VMM must have access
to certain registers in EL1, the ex_regs() flags gained the option to
run a thread in EL1. Unless one of the arch ex_regs flags is given, the
exception level is left unchanged. This makes the mechanism transparent
to most of the existing code.

On arm32, syscalls will just work, regardless of the chosen EL. In case
of a syscall from EL0 on cpu_virt, the "hvc" instruction will cause an
"undefined instruction" exception. The handler already checks for the
address in the syscall page and will recognize it as a syscall
nonetheless.

For arm64 it is the responsibility of the user space to utilize the
right instruction. On the KIP there will always be the "svc"
instruction because all user space is executed on EL0 by default. If a
program chooses to use EL1, it must make sure all syscalls use hvc and
that the KIP syscall entry is not used.

Change-Id: I2b949d873d541dbe16ebb124e53a258903030dcc