arm: always map kernel to constant virtual address
Remove the case of cpu_virt where the kernel virtual address was
identical to the physical address. Instead, the arm kernel will always
run at a fixed virtual address. Because we want the kernel image to be
relocatable in physical memory in the future, everything will be
consistently mapped by regular 1 or 2 MiB super pages.

The bootstrap code will only map the kernel image to the desired
virtual address. Likewise, only the kernel is added to the pmem map.
This implies that the free map can be taken as-is in
Kmem_alloc::Kmem_alloc() because the kernel is always a reserved memory
region in the Kip. In case of 32bit cpu_virt the 1:1 system RAM mapping
is still retained because it is later needed for the cache maintenance
functions in Mem_op::arm_mem_cache_maint().

Co-authored-by: Georg Kotheimer <georg.kotheimer@kernkonzept.com>
Change-Id: Ia1206863030b0ebef51a629deb3cc95a5bf0fb25