arm: prevent double mapping of pmem pages
At least the first 4 MiB of the RAM, containing the kernel image, are
mapped to Mem_layout::Map_base by the Fiasco bootstrap stage. This
mapping is added unconditionally to the Mem_layout::pmem translation
map. Any physical page that is recorded in the pmem map must be added
only once to have an unambiguous phys to virt (and vice versa) mapping.

Pages allocated for the kernel heap are mapped continuously in a
dedicated area in the kernel virtual memory space, starting from
Mem_layout::Pmem_start. These mappings are added to the
Mem_layout::pmem list too. Thus the heap page allocation must not use
any page that is already accounted in the Mem_layout::pmem map.
Otherwise they will be accounted twice with different virtual
addresses, which violates the required property that there is a unique
mapping between physical and virtual memory for all pmem pages.

Co-authored-by: Marcus Haehnel <marcus.haehnel@kernkonzept.com>
Change-Id: Id2373027a785ce97b9a94337f2566eb905a687c4