arm/smmuv3: Protect item complete check against wrap-around
Queue now stores the write queue index non-truncated, which makes it
possible to correctly determine the completion status of items that have
been overwritten.

In addition, use an Unsigned64 to represent write queue indexes, so that
it becomes effectively impossible for the index to wrap around and catch
up again while waiting for a command to complete.

See ticket #CD-944.

Change-Id: Iee5ea08a1be5a40e36482d895e5350e1ff25d6cd