arm: Move Linux cache maintenance API handling into function
The Linux cache maintenance API is only available on arm32. Emphasize
this by moving the implementation to a separate function in an arm32
source file. Furthermore, that makes it possible to include the
arm_mem_cache_maint mechanism, on which the Linux cache maintenance API
builds upon, only when compiling for arm32.

Change-Id: I12167022bafac4e8cac275da86a4561930f13acc