irq_mgr_multi_chip: Simplify IRQ number mapping
Using a bitmask to map global IRQ numbers to chip relative IRQ numbers
seems unnecessarily complicated without a real benefit, as the task of
Irq_mgr_multi_chip is only to determine the chip responsible for a
certain IRQ number range (with a granularity of Bits_per_entry).
The check whether a IRQ number is really valid is then done later in
the responsible IRQ chip.

Apart from that, the mask would be inadequate for such a check anyway,
that is, for the case where the number of IRQs a chip supports is not
a power of two.

Therefore, instead of a bitmask, just store the IRQ base for each chip.
While at it, also introduce a constant for the number of IRQs per entry.

Change-Id: Ie3d01aedd0bd8eaa09b9e441ca402c4174c96555