arm32: Ensure data operand on cache ops is zero
When invalidating the entire cache and performing some other cache
maintenance operations that do not require a data operand the generic
armv6 manual specifies that the data operand should be zero. Adhere to
this in all generic code to be conformat to the specification.

Reference manuals of some specific implementations (strong arm, pxa
or 926) may specify that the value is ignored instead. For
implementations specific to these platforms we thus leave the
corresponding implementations as is.

Change-Id: If76e13ae1c26765d778585df8cf38b2b2acaf1a3