arm32: use right TLB instructions on uniprocessor cpu_virt
The type of required TLB maintenance instructions is primarily
dependent on the exception level on which Fiasco is executing. When
running in EL2, we must always use the appropriate Virtualization
Extension instructions. Even if the CPU is single core, the IS (inner
shareable) variants of all TLB instructions can be used because an
implementation that includes the Virtualization Extension must also
include the Multiprocessing Extension.

For uniprocessor Armv7 configurations we must still assume that the CPU
does not sport the Virtualization Extension. Only starting with Armv8
we can default to the IS-variants of the EL1 TLB instructions.

See ticket #CD-1037.

Change-Id: I85307e36c1c9b8c27debd4edd418d561a563ea70