Guard against accessing static objects prior to construction
Objects declared using the Static_object<> template are prone to being
accidentally accessed (explicitly using the get() method or implicitly
using the * or the -> operators) prior to being properly constructed by
explicitly calling the construct() method.

If such objects contain non-trivial class members that require their
constructors to be also executed before use, such accidental accesses
prior to construction might constitute undefined behavior and can lead
to hard-to-debug bugs due to the downstream constructors being skipped.

This patch extends the Static_object<> template by a mechanism to
verify that the construct() method has been executed before any get()
method on an instance (and it also guards against calling the
construct() method multiple times).

To avoid having different object sizes in debug and non-debug builds,
the metadata is always present and updated (thus it can be examined
post mortem), but the assertions are no-ops in non-debug builds.

The metadata is initially set to false thanks to the fact that C++
guarantees that scalar members of objects are zero-initialized.

Also to avoid a potential infinite recursion in assert_fail() in
combination with the Kconsole singleton, a special non-checking variant
of the get() method is provided.

Note that for the sake of simplicity, the check against calling the
construct() method multiple times is not atomic. Therefore it cannot
reliably guard against very close races when called in parallel from
multiple threads.

Requires-Change-Id: I1bae00f5ff49d112d1785897b7594ce39ee4320b
Change-Id: I976c8566518d7df4dbeb871f0d94d259e5102422