arm: ensure observability of new PTE before returning
The Arm architecture does not cache entries that generate a translation
fault in the TLB. It is thus not necessary to flush the TLB for newly
mapped pages. But because the MMU is an independent observer, the write
of the new page table entry has to complete before returning to user
space. Otherwise the new entry might not be seen by the MMU, generating
another page fault, even though the table entry was already written.

Change-Id: Ia9f79e878cc1669091c2ee014274a00d1b981ed3