arm: ensure observability before writing table descriptor
The Arm architecture defines the MMU as a separate observer in the
memory model. The page table code thus must make sure that
(speculative) page table walks always see consistent data structures.

When we allocate new parts of the page table, we first set all entries
of the new table as invalid. Then the table descriptor in the
previous-level table is updated to point to the newly allocated chunk.
Because writes of the CPU are not necessarily observed in the same
order by the MMU, a memory barrier is necessary. This barriers ensures
that the clearing of the newly allocated table is observable before the
updated table descriptor.

Note that the Arm architecture defines that no "negative" hits are
cached in the TLB. That is, entries that generate a translation-, size-
or permission-faults are not cached. Therefore, Fiasco does not flush
the TLB when pages are mapped initially.

Change-Id: I87c2b58929338cb5cbe40d5368a58645de66f4a3